Renamed location_state.dart to old_location_state.dart to delete after refactoring and renamed add_location_state.dart to location_state.dart
