+++ lib/states/location_state.dart
@@ -4,8 +4,9 @@ import "package:photos/models/typedefs.dart";
 import "package:photos/utils/debouncer.dart";
 
 class LocationTagStateProvider extends StatefulWidget {
+  final List<double> centerPoint;
   final Widget child;
-  const LocationTagStateProvider(this.child, {super.key});
+  const LocationTagStateProvider(this.centerPoint, this.child, {super.key});
 
   @override
   State<LocationTagStateProvider> createState() =>
@@ -14,8 +15,14 @@ class LocationTagStateProvider extends StatefulWidget {
 
 class _LocationTagStateProviderState extends State<LocationTagStateProvider> {
   int selectedRaduisIndex = defaultRadiusValueIndex;
+  late List<double> centerPoint;
   final Debouncer _selectedRadiusDebouncer =
       Debouncer(const Duration(milliseconds: 300));
+  @override
+  void initState() {
+    centerPoint = widget.centerPoint;
+    super.initState();
+  }
 
   void _updateSelectedIndex(int index) {
     _selectedRadiusDebouncer.cancelDebounce();
@@ -32,6 +39,7 @@ class _LocationTagStateProviderState extends State<LocationTagStateProvider> {
   Widget build(BuildContext context) {
     return InheritedLocationTagData(
       selectedRaduisIndex,
+      centerPoint,
       _updateSelectedIndex,
       child: widget.child,
     );
@@ -40,9 +48,11 @@ class _LocationTagStateProviderState extends State<LocationTagStateProvider> {
 
 class InheritedLocationTagData extends InheritedWidget {
   final int selectedRadiusIndex;
+  final List<double> coordinates;
   final VoidCallbackParamInt updateSelectedIndex;
   const InheritedLocationTagData(
     this.selectedRadiusIndex,
+    this.coordinates,
     this.updateSelectedIndex, {
     required super.child,
     super.key,
