+++ lib/ui/viewer/location/add_location_sheet.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import "package:modal_bottom_sheet/modal_bottom_sheet.dart";
 import "package:photos/core/constants.dart";
 import "package:photos/services/location_service.dart";
-import "package:photos/states/add_location_state.dart";
+import 'package:photos/states/location_state.dart';
 import "package:photos/theme/colors.dart";
 import "package:photos/theme/ente_theme.dart";
 import "package:photos/ui/common/loading_widget.dart";
